test(site/src/pages/AgentsPage): port tool story contracts to Vitest and strip visual assertions
- Add toolShowcaseCoverage data test asserting every registered tool
  renderer has a showcase fixture; move allToolShowcaseItems to a shared
  toolShowcaseFixtures module imported by both the test and Tool stories.
- Delete the PolicyBadgeCoversEveryRenderer play (fixture completeness
  now covered by the data test; badge visibility is captured by Pixel).
- Remove the orphaned toolRendererNames export from Tool.tsx.
- Port AskUserQuestionTool payload contracts (non-default selection,
  Other-prefix formatting, multi-question numbering) into
  AskUserQuestionTool.test.tsx.
- Port ProposePlanTool implement and clipboard contracts into the new
  ProposePlanTool.test.tsx.
- Port the ToolCall accessible-name flip into the new ToolCall.test.tsx.
- Strip visual assertion lines from the corresponding plays; keep setup
  driving so Pixel captures the submitted/expanded states.
